backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pdfplumber==0.10.3
boto3==1.28.44
botocore==1.31.44
python-dotenv==1.0.0
httpx==0.25.2
anyio==3.7.1